Move TransformerCommon into the weco namespace
