+++ pipeline/transformer/transformer_common/src/main/scala/weco/catalogue/transformer/Transformer.scala
@@ -1,8 +1,8 @@
-package uk.ac.wellcome.transformer.common.worker
+package weco.catalogue.transformer
 
-import uk.ac.wellcome.models.work.internal._
+import uk.ac.wellcome.models.work.internal.Work
+import uk.ac.wellcome.models.work.internal.WorkState.Source
 import uk.ac.wellcome.models.work.internal.result.Result
-import WorkState.Source
 
 trait Transformer[SourceData] {
   def apply(sourceData: SourceData, version: Int): Result[Work[Source]]
