+++ pipeline/transformer/transformer_mets/src/main/scala/uk/ac/wellcome/platform/transformer/mets/transformer/MetsXmlTransformer.scala
@@ -5,8 +5,8 @@ import uk.ac.wellcome.storage.Identified
 import uk.ac.wellcome.models.work.internal.{Work, WorkState}
 import uk.ac.wellcome.models.work.internal.result.Result
 import uk.ac.wellcome.storage.s3.S3ObjectLocation
-import uk.ac.wellcome.transformer.common.worker.Transformer
 import weco.catalogue.source_model.mets.MetsSourceData
+import weco.catalogue.transformer.Transformer
 
 class MetsXmlTransformer(store: Readable[S3ObjectLocation, String])
     extends Transformer[MetsSourceData] {
